% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{time_features}
\alias{time_features}
\title{time features example: IBM, AAPL, AMZN, GOOGL and MSFT Close Prices}
\format{
A data frame with 6 columns and 1336 rows.
}
\source{
finance.yahoo.com
}
\usage{
time_features
}
\description{
A data frame with with daily with daily prices for some Big Tech Companies since March 2017.
}
\keyword{datasets}
